removed duplicate queryClient provider & unnecessary wrappers
